1.8 Port
Bumped version to 1.6.0

Renamed "mod_usefulfood" to "UsefulFood."

Replaced most (if not all) tab characters with spaces.
Added proper (from what I think) localization to tooltips and also to
the creative tab.

The bug was that the cakes were not properly mapped to each-other. They
were mapped individually and caused the texture system to do weird
stuff.

tl;dr
Items and blocks were not mapped to each-other, and that issue caused a
weird bug.

The bug fix may break worlds.